Mise en place de l'application Profile Refactor des tests et fichiers en conséquence
