+++ oc_lettings_site/admin.py
@@ -1,4 +1,4 @@
-from django.contrib import admin
-from .models import Profile
-
-admin.site.register(Profile)
+# from django.contrib import admin
+# from .profiles.models import Profile
+#
+# admin.site.register(Profile)
