+++ oc_lettings_site/settings.py
@@ -22,6 +22,7 @@ ALLOWED_HOSTS = ["*"]
 INSTALLED_APPS = [
     "oc_lettings_site.apps.OCLettingsSiteConfig",
     "oc_lettings_site.lettings.apps.LettingsConfig",
+    "oc_lettings_site.profiles.apps.ProfilesConfig",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
