+++ oc_lettings_site/urls.py
@@ -2,6 +2,7 @@ from django.contrib import admin
 from django.urls import path
 
 import oc_lettings_site.lettings.views
+import oc_lettings_site.profiles.views
 from . import views
 
 urlpatterns = [
@@ -10,8 +11,8 @@ urlpatterns = [
     path("lettings/", oc_lettings_site.lettings.views.index, name="lettings_index"),
     path("lettings/<str:letting_id>/", oc_lettings_site.lettings.views.letting, name="letting"),
     # Profile List/Select
-    path("profiles/", views.profiles_index, name="profiles_index"),
-    path("profiles/<str:username>/", views.profile, name="profile"),
+    path("profiles/", oc_lettings_site.profiles.views.index, name="profiles_index"),
+    path("profiles/<str:username>/", oc_lettings_site.profiles.views.profile, name="profile"),
     # Administration
     path("admin/", admin.site.urls),
 ]
